style: 작품 만족도 모달 뷰 로직 분리
