+++ src/constants/art/rate.ts
@@ -1,0 +1,1 @@
+export const TOTAL_STARS = 5;
