+++ src/pages/Art/ArtRate/types.ts
@@ -1,0 +1,8 @@
+export interface IArtRateProps {
+  isModalOpen: boolean;
+  showModal: () => void;
+  handleOk: () => void;
+  handleCancel: () => void;
+  FillStars: JSX.Element[];
+  EmptyStars: JSX.Element[];
+}
